+++ Entity/ApprovalStatus.php
@@ -18,8 +18,8 @@ use KimaiPlugin\ApprovalBundle\Repository\ApprovalStatusRepository;
  */
 class ApprovalStatus
 {
-    public const SUBMITTED = 'submitted'; //WRONG - only for migration
-    public const GRANTED = 'granted';
+    public const SUBMITTED = 'submitted'; 
+    public const GRANTED = 'granted'; //WRONG - only for migration
     public const DENIED = 'denied';
     public const APPROVED = 'approved';
     public const NOT_SUBMITTED = 'not_submitted';