+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 0.9 to 0.9.1
 
 - enhance tooltips
+- when a week is set to no approval, all later weeks which had been submitted or approved are reset to open